requirements.txt
celery==5.2.7
redis==4.5.1
snowflake-connector-python